remove batching regex from BatchRequest and BatchDefinition
