+++ great_expectations/datasource/fluent/invalid_datasource.py
@@ -27,8 +27,10 @@ from great_expectations.datasource.fluent.type_lookup import TypeLookup, ValidTy
 if TYPE_CHECKING:
     from great_expectations.core.partitioners import Partitioner
     from great_expectations.datasource.fluent.batch_request import BatchRequest
-    from great_expectations.datasource.fluent.interfaces import Batch, PartitionerProtocol
-
+    from great_expectations.datasource.fluent.interfaces import (
+        Batch,
+        PartitionerI,
+    )
 
 # Controls which methods should raise an error when called on an InvalidDatasource
 METHOD_SHOULD_RAISE_ERROR: Final[set] = {
@@ -88,7 +90,6 @@ class InvalidAsset(DataAsset):
         options: dict | None = None,
         batch_slice: Any = None,
         partitioner: Any = None,
-        batching_regex: Any = None,
     ) -> NoReturn:
         self._raise_type_error()
 
@@ -97,7 +98,7 @@ class InvalidAsset(DataAsset):
         self._raise_type_error()
 
     @override
-    def sort_batches(self, batch_list: List[Batch], partitioner: PartitionerProtocol) -> None:
+    def sort_batches(self, batch_list: List[Batch], partitioner: PartitionerI) -> None:
         self._raise_type_error()
 
     @override
