+++ great_expectations/datasource/fluent/pandas_datasource.pyi
@@ -1,5 +1,4 @@
 import os
-import re
 import sqlite3
 import typing
 from logging import Logger
@@ -79,7 +78,6 @@ class _PandasDataAsset(DataAsset):
         options: Optional[BatchParameters] = ...,
         batch_slice: Optional[BatchSlice] = ...,
         partitioner: Optional[Partitioner] = ...,
-        batching_regex: Optional[re.Pattern] = ...,
     ) -> BatchRequest: ...
     @override
     def _validate_batch_request(self, batch_request: BatchRequest) -> None: ...
