+++ great_expectations/datasource/fluent/spark_datasource.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import logging
-import re
 import warnings
 from pprint import pformat as pf
 from typing import (
@@ -208,13 +207,12 @@ class DataFrameAsset(DataAsset, Generic[_SparkDataFrameT]):
         version="0.16.15",
     )
     @override
-    def build_batch_request(  # type: ignore[override]   # noqa: PLR0913
+    def build_batch_request(  # type: ignore[override]
         self,
         dataframe: Optional[_SparkDataFrameT] = None,
         options: Optional[BatchParameters] = None,
         batch_slice: Optional[BatchSlice] = None,
         partitioner: Optional[Partitioner] = None,
-        batching_regex: Optional[re.Pattern] = None,
     ) -> BatchRequest:
         """A batch request that can be used to obtain batches for this DataAsset.
 
@@ -223,7 +221,6 @@ class DataFrameAsset(DataAsset, Generic[_SparkDataFrameT]):
             options: This is not currently supported and must be {}/None for this data asset.
             batch_slice: This is not currently supported and must be None for this data asset.
             partitioner: This is not currently supported and must be None for this data asset.
-            batching_regex: This is currently not supported and must be None for this data asset.
 
 
         Returns:
@@ -243,10 +240,6 @@ class DataFrameAsset(DataAsset, Generic[_SparkDataFrameT]):
         if partitioner is not None:
             raise ValueError(  # noqa: TRY003
                 "partitioner is not currently supported and must be None for this DataAsset."
-            )
-        if batching_regex is not None:
-            raise ValueError(  # noqa: TRY003
-                "batching_regex is not currently supported and must be None for this DataAsset."
             )
 
         if dataframe is None:
