+++ tests/datasource/fluent/test_metadatasource.py
@@ -4,7 +4,6 @@ import copy
 import inspect
 import logging
 import pathlib
-import re
 from pprint import pformat as pf
 from typing import TYPE_CHECKING, ClassVar, Dict, Generator, List, Optional, Tuple, Type, Union
 
@@ -126,7 +125,6 @@ class DummyDataAsset(DataAsset):
         options: Optional[BatchParameters] = None,
         batch_slice: Optional[BatchSlice] = None,
         partitioner: Optional[Partitioner] = None,
-        batching_regex: Optional[re.Pattern] = None,
     ) -> BatchRequest:
         return BatchRequest("datasource_name", "data_asset_name", options or {})
